Implement fetch comics
This commit implements fetching the comics from XKCD and printing them to the console.

It would later on publish them to a channel that would be used by a different usecase to process the into elastic search.

The current implementation doesn't store the latest fetched comic in any persistent way, but in the future it can be stored by a database (e.g. Postgres), so comics can be replayed if needed without spamming the API of XKCD